// src/components/BiometricLogin/generated.ts
/*********** Manifest Props *******************
 * This file is auto generated, making manual *
 * edits to this file might result in loosing *
 * information.                               *
 **********************************************/
export interface IStyles {
  fontFamily?: string
  fontSize?: number
  fontWeight?: "normal" | "bold" | "100" | "200" | "300" | "400" | "500" | "600" | "700" | "800" | "900" | undefined
  textAlignment?: string
  color?: string
}

export interface IFonts {
  body: string
  heading: string
}

export interface IAvatar {
  uri: string
  cache: string
}


export interface BiometricLoginProps {
  title?: string
  styles: { title: IStyles }
  icon?: string
  backgroundColor?: string
  username?: string
  password?: string
  validCredentials?: string
  onCredentialRetriveSuccess?: (Username?: string, Password?: string) => void
  onCredentialRetriveFailed?: (Errormessage?: string) => void
  onEnrollmentSuccess?: () => void
  onEnrollmentFailure?: () => void
  appId: string
  _fonts: IFonts
  _width: number
  _height: number
  editor: boolean
}